Move types out of __init__.py file in types subfolder into a custom_types.py file
