+++ src/lambdas/determine_replacements_caselaw/index.py
@@ -9,9 +9,9 @@ from aws_lambda_powertools.utilities.data_classes import S3Event, event_source
 from aws_lambda_powertools.utilities.data_classes.s3_event import S3EventRecord
 from aws_lambda_powertools.utilities.typing import LambdaContext
 
+from utils.custom_types import DocumentAsXMLString
 from utils.environment_helpers import validate_env_variable
 from utils.initialise_db import init_db_connection
-from utils.types import DocumentAsXMLString
 
 if TYPE_CHECKING:
     from mypy_boto3_sqs.type_defs import MessageAttributeValueTypeDef
