+++ src/lambdas/extract_judgement_contents/index.py
@@ -7,9 +7,9 @@ from aws_lambda_powertools.utilities.data_classes import S3Event, event_source
 from aws_lambda_powertools.utilities.data_classes.s3_event import S3EventRecord
 from aws_lambda_powertools.utilities.typing import LambdaContext
 
+from utils.custom_types import DocumentAsXMLString
 from utils.environment_helpers import validate_env_variable
 from utils.helper import parse_file
-from utils.types import DocumentAsXMLString
 
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.INFO)
