+++ src/lambdas/fetch_xml/index.py
@@ -8,8 +8,8 @@ from aws_lambda_powertools.utilities.data_classes import SQSEvent, event_source
 from aws_lambda_powertools.utilities.data_classes.sqs_event import SQSRecord
 from aws_lambda_powertools.utilities.typing import LambdaContext
 
+from utils.custom_types import APIEndpointBaseURL, DocumentAsXMLString
 from utils.environment_helpers import validate_env_variable
-from utils.types import APIEndpointBaseURL, DocumentAsXMLString
 
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.INFO)
