+++ src/lambdas/make_replacements/index.py
@@ -8,8 +8,8 @@ from aws_lambda_powertools.utilities.data_classes.sqs_event import SQSRecord
 from aws_lambda_powertools.utilities.typing import LambdaContext
 
 from replacer.make_replacments import make_post_header_replacements
+from utils.custom_types import DocumentAsXMLString
 from utils.environment_helpers import validate_env_variable
-from utils.types import DocumentAsXMLString
 
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.DEBUG)
