+++ src/lambdas/push_enriched_xml/index.py
@@ -9,8 +9,8 @@ from aws_lambda_powertools.utilities.data_classes.sqs_event import SQSRecord
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from requests.auth import HTTPBasicAuth
 
+from utils.custom_types import APIEndpointBaseURL, DocumentAsXMLString
 from utils.environment_helpers import validate_env_variable
-from utils.types import APIEndpointBaseURL, DocumentAsXMLString
 
 LOGGER = logging.getLogger()
 LOGGER.setLevel(logging.INFO)
