+++ src/lambdas/xml_validate/index.py
@@ -10,8 +10,8 @@ from aws_lambda_powertools.utilities.data_classes.s3_event import S3EventRecord
 from aws_lambda_powertools.utilities.typing import LambdaContext
 from lxml import etree
 
+from utils.custom_types import DocumentAsXMLBytes
 from utils.environment_helpers import validate_env_variable
-from utils.types import DocumentAsXMLBytes
 
 if TYPE_CHECKING:
     from mypy_boto3_sqs.type_defs import MessageAttributeValueTypeDef
